fixed voting error message
