+++ frontend-dart/lib/core/builder_core.dart
@@ -2,7 +2,7 @@ import 'navigation_core.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:provider/single_child_widget.dart';
-import '../core/constants_core.dart';
+import '../core/provider_core.dart';
 import '../core/locator_core.dart';
 import '../widgets/app_widgets.dart';
 import '../providers/auth_providers.dart';
@@ -25,7 +25,6 @@ import '../screens/playlists/editor_playlists.dart';
 import '../screens/playlists/sharing_playlists.dart';
 import '../screens/playlists/main_playlists.dart';
 import '../screens/profile/profile_screens.dart';
-import '../screens/profile/social_profile.dart';
 import '../screens/profile/password_profile.dart';
 import '../screens/profile/user_profile.dart';
 import '../screens/friends/add_friends.dart';
