+++ frontend-dart/lib/main.dart
@@ -5,7 +5,7 @@ import 'package:provider/provider.dart';
 import 'package:responsive_framework/responsive_framework.dart';
 import 'core/theme_core.dart';
 import 'core/social_core.dart';
-import 'core/constants_core.dart';
+import 'core/provider_core.dart';
 import 'core/locator_core.dart';
 import 'core/builder_core.dart';
 import 'core/navigation_core.dart';
