+++ frontend-dart/lib/providers/voting_providers.dart
@@ -219,7 +219,28 @@ class VotingProvider extends BaseProvider {
         AppLogger.error('Invalid track index sent to backend: $trackIndex', 'VotingProvider');
       } else {
         AppLogger.error('Vote failed with error', e, null, 'VotingProvider');
-        setError('Voting failed. Please try again.');
+        
+        // Extract the actual error message from the backend response
+        String errorMessage = 'Voting failed. Please try again.';
+        
+        if (e is DioException && e.response?.data != null) {
+          try {
+            final responseData = e.response!.data;
+            if (responseData is Map<String, dynamic>) {
+              // Try to get 'detail' field first, then 'message', then 'error'
+              errorMessage = responseData['detail']?.toString() ?? 
+                             responseData['message']?.toString() ?? 
+                             responseData['error']?.toString() ?? 
+                             errorMessage;
+            } else if (responseData is String) {
+              errorMessage = responseData;
+            }
+          } catch (extractError) {
+            AppLogger.debug('[VoteForTrackByIndex] Failed to extract error message: $extractError', 'VotingProvider');
+          }
+        }
+        
+        setError(errorMessage);
       }
       
       return false;
