+++ frontend-dart/lib/screens/auth/auth_screens.dart
@@ -4,7 +4,7 @@ import 'package:provider/provider.dart';
 import '../../providers/auth_providers.dart';
 import '../../core/theme_core.dart';
 import '../../core/responsive_core.dart';
-import '../../core/constants_core.dart';
+import '../../core/provider_core.dart';
 import '../../core/logging_core.dart';
 import '../../widgets/app_widgets.dart';
 import '../base_screens.dart';
