+++ frontend-dart/lib/screens/home/home_screens.dart
@@ -9,11 +9,11 @@ import '../../services/player_services.dart';
 import '../../core/locator_core.dart';
 import '../../core/theme_core.dart';
 import '../../core/responsive_core.dart';
-import '../../core/constants_core.dart';
+import '../../core/provider_core.dart';
 import '../../core/logging_core.dart';
 import '../../core/animations_core.dart';
 import '../../widgets/app_widgets.dart';
-import '../../widgets/status_widgets.dart';
+import '../../widgets/state_widgets.dart';
 import '../../models/music_models.dart';
 import '../profile/profile_screens.dart';
 import '../music/search_music.dart';
