+++ frontend-dart/lib/screens/profile/profile_screens.dart
@@ -3,9 +3,9 @@ import 'package:flutter/foundation.dart' show kDebugMode, debugPrint;
 import 'package:provider/provider.dart';
 import '../../providers/profile_providers.dart';
 import '../../core/theme_core.dart';
-import '../../core/constants_core.dart';
+import '../../core/provider_core.dart';
 import '../../core/logging_core.dart';
-import '../../widgets/status_widgets.dart';
+import '../../widgets/state_widgets.dart';
 import '../../widgets/avatar_widgets.dart';
 import '../../widgets/sections_widgets.dart';
 import '../base_screens.dart';
