+++ frontend-dart/test/services/activity_services_test.dart
@@ -1,7 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:mockito/annotations.dart';
-import 'package:music_room/services/activity_services.dart';
 import 'package:music_room/services/api_services.dart';
 import 'package:music_room/providers/auth_providers.dart';
 import 'package:music_room/models/api_models.dart';
